# points-service/swagger.yaml
openapi: 3.0.0
info:
  title: Points Service API
  description: Points Service to manage spins, points, and coins balance for a slot machine game.
  version: 1.0.0
servers:
  - url: http://localhost:3001
paths:
  /balance/{userId}:
    get:
      summary: Get the balance of spins, points, and coins
      tags:
        - Points
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
            description: The ID of the user
      responses:
        '200':
          description: Successful response with spins, points, and coins
          content:
            application/json:
              schema:
                type: object
                properties:
                  spins:
                    type: integer
                  points:
                    type: integer
                  coins:
                    type: integer
              example:
                spins: 5
                points: 100
                coins: 10
        '500':
          description: Internal server error
  /balance/{userId}/{resource}:
    get:
      summary: Get the balance of a specific resource (spins, points, or coins)
      tags:
        - Points
      parameters:
        - in: path
          name: userId
          required: true
          schema:
            type: string
        - in: path
          name: resource
          required: true
          schema:
            type: string
            enum: [points, spins, coins]
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: object
                properties:
                  balance:
                    type: integer
              example:
                balance: 100
        '500':
          description: Internal server error
  /update:
    post:
      summary: Update the user's points, coins, and spins after a spin
      tags:
        - Points
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                spinResult:
                  type: array
                  items:
                    type: integer
                spinsUsed:
                  type: integer
            example:
              userId: "user123"
              spinResult: [7, 7, 7]
              spinsUsed: 1
      responses:
        '200':
          description: Successfully updated
          content:
            application/json:
              schema:
                type: object
                properties:
                  pointsEarned:
                    type: integer
                  newBalance:
                    type: integer
                  newCoinsBalance:
                    type: integer
                  missionCompleted:
                    type: boolean
              example:
                pointsEarned: 21
                newBalance: 71
                newCoinsBalance: 10
                missionCompleted: true
        '500':
          description: Internal server error
  /set-balance:
    post:
      summary: Set the balance of spins, points, or coins for a user
      tags:
        - Points
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                userId:
                  type: string
                resource:
                  type: string
                  enum: [ points, spins, coins ]
                amount:
                  type: integer
            example:
              userId: "user123"
              resource: "spins"
              amount: 100  # Example to set spins balance
      responses:
        '200':
          description: Balance successfully set
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  newBalance:
                    type: integer
                example:
                  message: "Spins balance updated successfully"
                  newBalance: 100
        '500':
          description: Internal server error